Template de Feria especifica
-Se agrega template para feria especifica
-Se elimina función redundante para calcular distancias
-Se agrega función para conseguir Slug en modelo Feria
