+++ api/views/utils.py
@@ -2,34 +2,8 @@
 # Clases encargadas que ayudan a las vistas
 #
 # Author: Tyron Fonseca - tyron.fonseca@ucr.ac.cr
-# Last modified: 29/11/2021
+# Last modified: 4/1/2022
 # ===================================================
-
-import math
-
-
-class FeriasHelper:
-    def is_in_radius(lat1, lon1, lat2, lon2, radius):
-        """ Calcular distancia de dos coordenadas usando la formula Haversine
-            https://www.movable-type.co.uk/scripts/latlong.html
-        """
-
-        radius_earth = 6371e3  # Radio de la tierra en metros
-        pi_radian = (math.pi/180)
-        phi_1 = lat1 * pi_radian
-        phi_2 = lat2 * pi_radian
-        delta_phi = (lat2 - lat1) * pi_radian
-        delta_lambda = (lon2 - lon1) * pi_radian
-
-        a = math.sin(delta_phi/2)**2 + math.cos(phi_1) * \
-            math.cos(phi_2) * math.sin(delta_lambda/2) ** 2
-        c = 2 * math.atan2(math.sqrt(a), math.sqrt(1-a))
-        distance = radius_earth * c
-        if distance <= radius:
-            return True
-        else:
-            return False
-
 
 class DynamicFieldsViewMixin(object):
     """
