+++ ferias/templates/ferias_detail.html
@@ -1,0 +1,15 @@
+{% extends "base.html" %}
+{% load static %}
+
+{% block title %}{{feria.nombre}}{% endblock title %}
+
+{% block styles %}
+{% endblock %}
+{% block scripts %}
+{% endblock %}
+
+{% block content %}
+<h1>{{feria.nombre}}</h1>
+{% endblock %}
+{% block scriptsEnd %}
+{% endblock %}
