+++ ferias/urls.py
@@ -4,4 +4,8 @@ from . import views
 
 urlpatterns = [
     path('', views.ferias, name='list-ferias'),
+    path('<str:feria_id>/',
+         views.ferias_detail, name='feria-detail'),
+    path('<str:feria_id>/<slug:slug>/', 
+         views.ferias_detail, name='feria-detail'),
 ]
